pyproject.toml
[tool.poetry]
name = "reach_beta"
version = "0.1.0"
description = ""
authors = ["mpdl <153767344+mapodele@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
crewai = "^0.51.1"
pydantic = "^2.8.2"
openai = "^1.43.0"
pyyaml = "^6.0.2"
crewai-tools = "^0.12.0"
google-ads = "^25.0.0"
pillow = "^10.4.0"
scipy = "^1.14.1"
matplotlib = "^3.9.2"
webdriver-manager = "^4.0.2"
gspread = "^6.1.2"
gspread-dataframe = "^4.0.0"
tensorflow = "^2.17.0"
keras = "^3.5.0"
opencv-contrib-python-headless = "^4.10.0.84"
pytesseract = "^0.3.13"
selenium = "^4.24.0"
requests = "^2.32.3"
pandas = "^2.2.2"
exifread = "^3.0.0"
scikit-image = "^0.24.0"
tensorflow-hub = "^0.16.1"
tensorflow-object-detection-api = "^0.1.1"
ultralytics = "^8.2.98"
svgwrite = "^1.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
